Rts/elf-linker: Upcast to 64bit to satisfy format string.
The elf size is 32bit on 32bit builds and 64 otherwise.
We just upcast to 64bits before printing now.

(cherry picked from commit ed22678a7060a95f82804072b0e67ae4d09bf023)

+++ rts/linker/Elf.c
@@ -919,8 +919,8 @@ ocGetNames_ELF ( ObjectCode* oc )
                    common_used += symbol->elf_sym->st_size;
                    ASSERT(common_used <= common_size);
 
-                   debugBelch("COMMON symbol, size %ld name %s allocated at %p\n",
-                                   symbol->elf_sym->st_size, nm, symbol->addr);
+                   debugBelch("COMMON symbol, size %llu name %s allocated at %p\n",
+                                   (long long unsigned int) symbol->elf_sym->st_size, nm, symbol->addr);
 
                    /* Pointless to do addProddableBlock() for this area,
                       since the linker should never poke around in it. */
